refactor: improve easy table and utils methods
