+++ src/app/routes/dashboard/dashboard.component.ts
@@ -5,62 +5,9 @@ import {
   OnDestroy,
   ChangeDetectionStrategy,
 } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
 
 import { SettingsService } from '@core';
-
-export interface PeriodicElement {
-  name: string;
-  position: number;
-  weight: number;
-  symbol: string;
-}
-
-const ELEMENT_DATA: PeriodicElement[] = [
-  { position: 1, name: 'Hydrogen', weight: 1.0079, symbol: 'H' },
-  { position: 2, name: 'Helium', weight: 4.0026, symbol: 'He' },
-  { position: 3, name: 'Lithium', weight: 6.941, symbol: 'Li' },
-  { position: 4, name: 'Beryllium', weight: 9.0122, symbol: 'Be' },
-  { position: 5, name: 'Boron', weight: 10.811, symbol: 'B' },
-  { position: 6, name: 'Carbon', weight: 12.0107, symbol: 'C' },
-  { position: 7, name: 'Nitrogen', weight: 14.0067, symbol: 'N' },
-  { position: 8, name: 'Oxygen', weight: 15.9994, symbol: 'O' },
-  { position: 9, name: 'Fluorine', weight: 18.9984, symbol: 'F' },
-  { position: 10, name: 'Neon', weight: 20.1797, symbol: 'Ne' },
-];
-
-const MESSAGES = [
-  {
-    img: 'assets/images/avatars/avatar-1.jpg',
-    subject: 'Hydrogen',
-    content: `Cras sit amet nibh libero, in gravida nulla.
-     Nulla vel metus scelerisque ante sollicitudin commodo.`,
-  },
-  {
-    img: 'assets/images/avatars/avatar-2.jpg',
-    subject: 'Helium',
-    content: `Cras sit amet nibh libero, in gravida nulla.
-     Nulla vel metus scelerisque ante sollicitudin commodo.`,
-  },
-  {
-    img: 'assets/images/avatars/avatar-3.jpg',
-    subject: 'Lithium',
-    content: `Cras sit amet nibh libero, in gravida nulla.
-     Nulla vel metus scelerisque ante sollicitudin commodo.`,
-  },
-  {
-    img: 'assets/images/avatars/avatar-4.jpg',
-    subject: 'Beryllium',
-    content: `Cras sit amet nibh libero, in gravida nulla.
-     Nulla vel metus scelerisque ante sollicitudin commodo.`,
-  },
-  {
-    img: 'assets/images/avatars/avatar-6.jpg',
-    subject: 'Boron',
-    content: `Cras sit amet nibh libero, in gravida nulla.
-     Nulla vel metus scelerisque ante sollicitudin commodo.`,
-  },
-];
+import { DashboardService } from './dashboard.srevice';
 
 @Component({
   selector: 'app-dashboard',
@@ -74,17 +21,18 @@ const MESSAGES = [
     `,
   ],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [DashboardService],
 })
 export class DashboardComponent implements OnInit, AfterViewInit, OnDestroy {
   displayedColumns: string[] = ['position', 'name', 'weight', 'symbol'];
-  dataSource = ELEMENT_DATA;
+  dataSource = this.dashboardSrv.getData();
 
-  messages = MESSAGES;
+  messages = this.dashboardSrv.getMessages();
 
   chart1 = null;
   chart2 = null;
 
-  constructor(private http: HttpClient, private settings: SettingsService) {}
+  constructor(private settings: SettingsService, private dashboardSrv: DashboardService) {}
 
   ngOnInit() {}
