+++ src/app/routes/tables/basic/basic.component.html
@@ -1,6 +1,7 @@
 <page-header class="bg-green-600"></page-header>
 
-<easy-table [data]="list"
+<easy-table class="mat-elevation-z1"
+            [data]="list"
             [columns]="columns"
             [loading]="isLoading"
             [front]="true">
