+++ src/app/shared/components/easy-table/easy-table-cell.component.html
@@ -62,7 +62,7 @@
       <!-- {{(cellValue | percent) || '--'}} -->
       <!-- {{(cellValue | numberFixed | numberFormat:cell.type) || '--'}} -->
     </ng-container>
-    <!-- 普通 -->
+    <!-- 默认 -->
     <ng-container *ngSwitchDefault>
       {{cellValue || '--'}}
     </ng-container>
